Digit-Core/core-services/crop-disease-detection/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
numpy==1.24.3
Pillow==10.0.0
tensorflow==2.13.0
keras==2.13.1
opencv-python==4.8.0.76
scikit-learn==1.3.0
scipy==1.11.2
pandas==2.0.3
matplotlib==3.7.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
